A ready-made function to backup to ZIP file
A ready-made function to backup selected type of files within a folder.
Backup will be in ZIP file. Also, if program is run multiple times for a selected folders, then backup will exclude the previous backup ZIP files and rename the current backup file as filename_1.zip, filename_2.zip and so on.